Redirect authenticated users from index to dashboard
